backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
flasgger==0.9.7.1
Werkzeug==3.0.1
cryptography==41.0.7
pymongo==4.6.1
flask-pymongo==2.3.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
